mingit: switch from uname -m to MSYSTEM
$(uname -m) returns the architecture of the MSYS2 shell. This works for
x64 and x86 on Windows, but not for ARM64, for which we'll add support
in a follow-up PR. The reason is that MSYS2 can only be run using x64
emulation on ARM64 due to lacking Cygwin support for that platform.

To prepare for ARM64 support and to be more consistent with other MSYS2
logic, let's leverage $MSYSTEM instead of $(uname -m).

Ref: https://www.msys2.org/wiki/arm64/
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>