git-extra: set arch to "any"
Trying to install the `aarch64` package resulted in the following error:
"package git-extra-1.1.614.d551cf0cc-1-aarch64 does not have a valid
architecture".

This might be because MSYS2 is running on x64 emulation on Windows
ARM64. All of the other GfW packages have `any` as the architecture as
well, so let's use that here too. This ensures that the `aarch64`
package can be installed correctly.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>